requirements.txt
Cerberus==1.3.5
certifi==2024.6.2
charset-normalizer==3.3.2
evarify==0.1.1
future==1.0.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
python-dotenv==1.0.1
python-logging-loki==0.3.1
PyYAML==6.0.1
requests==2.32.3
rfc3339==6.2
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
